requirements.txt
django==5.0.3
asgiref
autopep8
django-cors-headers
djangorestframework
Pillow
pycodestyle
pytz
sqlparse
toml
python-dotenv==1.0.1
gunicorn==21.2.0
djangorestframework_simplejwt==5.3.1 
pyjwt==2.8.0